+++ src/docs/asciidoc/api/stock_admin.adoc
@@ -1,0 +1,13 @@
+[[stock-refresh]]
+=== 종목 최신화
+
+해당 API는 관리자 또는 매니저 권한이 요구됩니다.
+
+==== HTTP Request
+
+include::{snippets}/stock-refresh/http-request.adoc[]
+
+==== HTTP Response
+
+include::{snippets}/stock-refresh/http-response.adoc[]
+include::{snippets}/stock-refresh/response-fields.adoc[]
