+++ src/docs/asciidoc/api/stock_target_price_notification.adoc
@@ -1,0 +1,13 @@
+[[stock_target_price-notification-notify]]
+=== 종목 지정가 알림 발송
+
+해당 API는 관리자 또는 매니저 권한이 요구됩니다.
+
+==== HTTP Request
+
+include::{snippets}/stock_target_price-notification-notify/http-request.adoc[]
+
+==== HTTP Response
+
+include::{snippets}/stock_target_price-notification-notify/http-response.adoc[]
+include::{snippets}/stock_target_price-notification-notify/response-fields.adoc[]
