+++ src/main/java/codesquad/fineants/domain/stock_target_price/service/StockTargetPriceService.java
@@ -1,8 +1,6 @@
 package codesquad.fineants.domain.stock_target_price.service;
 
 import java.util.List;
-import java.util.concurrent.Executor;
-import java.util.concurrent.Executors;
 import java.util.stream.Collectors;
 
 import org.springframework.security.access.annotation.Secured;
@@ -18,7 +16,6 @@ import codesquad.fineants.domain.stock.repository.StockRepository;
 import codesquad.fineants.domain.stock_target_price.domain.dto.request.TargetPriceNotificationCreateRequest;
 import codesquad.fineants.domain.stock_target_price.domain.dto.request.TargetPriceNotificationUpdateRequest;
 import codesquad.fineants.domain.stock_target_price.domain.dto.response.TargetPriceNotificationCreateResponse;
-import codesquad.fineants.domain.stock_target_price.domain.dto.response.TargetPriceNotificationDeleteResponse;
 import codesquad.fineants.domain.stock_target_price.domain.dto.response.TargetPriceNotificationSearchItem;
 import codesquad.fineants.domain.stock_target_price.domain.dto.response.TargetPriceNotificationSearchResponse;
 import codesquad.fineants.domain.stock_target_price.domain.dto.response.TargetPriceNotificationSpecificItem;
@@ -30,9 +27,7 @@ import codesquad.fineants.domain.stock_target_price.repository.StockTargetPriceR
 import codesquad.fineants.domain.stock_target_price.repository.TargetPriceNotificationRepository;
 import codesquad.fineants.global.common.authorized.Authorized;
 import codesquad.fineants.global.common.authorized.service.StockTargetPriceAuthorizedService;
-import codesquad.fineants.global.common.authorized.service.TargetPriceNotificationAuthorizedService;
 import codesquad.fineants.global.common.resource.ResourceId;
-import codesquad.fineants.global.common.resource.ResourceIds;
 import codesquad.fineants.global.errors.errorcode.MemberErrorCode;
 import codesquad.fineants.global.errors.errorcode.StockErrorCode;
 import codesquad.fineants.global.errors.exception.BadRequestException;
@@ -44,13 +39,7 @@ import lombok.extern.slf4j.Slf4j;
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class StockTargetPriceNotificationService {
-
-	private static final Executor executor = Executors.newFixedThreadPool(100, r -> {
-		Thread thread = new Thread(r);
-		thread.setDaemon(true);
-		return thread;
-	});
+public class StockTargetPriceService {
 	private static final int TARGET_PRICE_NOTIFICATION_LIMIT = 5;
 
 	private final StockTargetPriceRepository repository;
@@ -59,9 +48,10 @@ public class StockTargetPriceNotificationService {
 	private final StockRepository stockRepository;
 	private final ClosingPriceRepository manager;
 
+	// 종목 지정가 및 지정가 알림 생성
 	@Transactional
 	@Secured("ROLE_USER")
-	public TargetPriceNotificationCreateResponse createStockTargetPriceNotification(
+	public TargetPriceNotificationCreateResponse createStockTargetPrice(
 		TargetPriceNotificationCreateRequest request,
 		Long memberId
 	) {
@@ -118,90 +108,8 @@ public class StockTargetPriceNotificationService {
 			.orElseThrow(() -> new NotFoundResourceException(StockErrorCode.NOT_FOUND_STOCK));
 	}
 
-	// 종목 지정가 알림 단일 제거
-	@Transactional
-	@Authorized(serviceClass = TargetPriceNotificationAuthorizedService.class)
 	@Secured("ROLE_USER")
-	public TargetPriceNotificationDeleteResponse deleteStockTargetPriceNotification(
-		@ResourceId Long targetPriceNotificationId
-	) {
-		List<Long> targetPriceNotificationIds = List.of(targetPriceNotificationId);
-		// 삭제하고자 하는 종목 지정가가 존재하는지 검증
-		verifyExistTargetPriceById(targetPriceNotificationIds);
-
-		TargetPriceNotification targetPriceNotification = findTargetPriceNotification(targetPriceNotificationId);
-		int deletedTargetPriceNotificationCount = deleteAllTargetPriceNotification(targetPriceNotificationIds);
-
-		// TargetPriceNotification이 한개도 없으면 StockTargetPrice 제거
-		if (targetPriceNotification.getStockTargetPrice().getTargetPriceNotifications().isEmpty()) {
-			repository.deleteById(targetPriceNotification.getStockTargetPrice().getId());
-		}
-
-		log.info("종목 지정가 알림 제거 결과 : targetPriceNotificationIds={}, deletedTargetPriceNotificationCount={}",
-			targetPriceNotificationIds, deletedTargetPriceNotificationCount);
-		return TargetPriceNotificationDeleteResponse.from(targetPriceNotificationIds);
-	}
-
-	private TargetPriceNotification findTargetPriceNotification(Long targetPriceNotificationId) {
-		return targetPriceNotificationRepository.findById(targetPriceNotificationId)
-			.orElseThrow(() -> new NotFoundResourceException(StockErrorCode.NOT_FOUND_TARGET_PRICE));
-	}
-
-	// 종목 지정가 존재 검증
-	private void verifyExistTargetPriceById(List<Long> targetPriceNotificationIds) {
-		if (targetPriceNotificationRepository.findAllById(targetPriceNotificationIds).size()
-			!= targetPriceNotificationIds.size()) {
-			throw new NotFoundResourceException(StockErrorCode.NOT_FOUND_TARGET_PRICE);
-		}
-	}
-
-	// 종목 지정가 단일 제거
-	@Transactional
-	@Authorized(serviceClass = StockTargetPriceAuthorizedService.class)
-	public void deleteStockTargetPrice(@ResourceId Long stockTargetPriceId) {
-		StockTargetPrice stockTargetPrice = repository.findById(stockTargetPriceId)
-			.orElseThrow(() -> new NotFoundResourceException(StockErrorCode.NOT_FOUND_STOCK_TARGET_PRICE));
-		targetPriceNotificationRepository.deleteAllByStockTargetPrices(List.of(stockTargetPrice));
-		repository.deleteById(stockTargetPriceId);
-	}
-
-	// 종목 지정가 알림 다수 제거
-	@Transactional
-	@Authorized(serviceClass = TargetPriceNotificationAuthorizedService.class)
-	@Secured("ROLE_USER")
-	public TargetPriceNotificationDeleteResponse deleteAllStockTargetPriceNotification(
-		@ResourceIds List<Long> targetPriceNotificationIds,
-		String tickerSymbol,
-		Long memberId
-	) {
-		// 존재하지 않는 종목 검증
-		verifyExistStock(tickerSymbol);
-		// 존재하지 않는 종목 지정가가 있는지 검증
-		verifyExistTargetPriceById(targetPriceNotificationIds);
-
-		int deletedCount = deleteAllTargetPriceNotification(targetPriceNotificationIds);
-		int deletedStockTargetPriceCount = deleteStockTargetPrice(tickerSymbol, memberId);
-		log.info("종목 지정가 알림 제거 결과 : ids={}, deletedTargetPriceNotificationCount={}, deletedStockTargetPriceCount = {}",
-			targetPriceNotificationIds, deletedCount, deletedStockTargetPriceCount);
-		return TargetPriceNotificationDeleteResponse.from(targetPriceNotificationIds);
-	}
-
-	private void verifyExistStock(String tickerSymbol) {
-		if (stockRepository.findByTickerSymbol(tickerSymbol).isEmpty()) {
-			throw new NotFoundResourceException(StockErrorCode.NOT_FOUND_STOCK);
-		}
-	}
-
-	private int deleteAllTargetPriceNotification(List<Long> targetPriceNotificationIds) {
-		return targetPriceNotificationRepository.deleteAllByTargetPriceNotificationIds(targetPriceNotificationIds);
-	}
-
-	private int deleteStockTargetPrice(String tickerSymbol, Long memberId) {
-		return repository.deleteByTickerSymbolAndMemberId(tickerSymbol, memberId);
-	}
-
-	@Secured("ROLE_USER")
-	public TargetPriceNotificationSearchResponse searchStockTargetPriceNotification(Long memberId) {
+	public TargetPriceNotificationSearchResponse searchStockTargetPrices(Long memberId) {
 		List<StockTargetPrice> stockTargetPrices = repository.findAllByMemberId(memberId);
 		List<TargetPriceNotificationSearchItem> stocks = stockTargetPrices.stream()
 			.map(stockTargetPrice -> TargetPriceNotificationSearchItem.from(stockTargetPrice, manager))
@@ -209,19 +117,8 @@ public class StockTargetPriceNotificationService {
 		return TargetPriceNotificationSearchResponse.from(stocks);
 	}
 
-	@Transactional
 	@Secured("ROLE_USER")
-	public TargetPriceNotificationUpdateResponse updateStockTargetPriceNotification(
-		TargetPriceNotificationUpdateRequest request, Long memberId) {
-		StockTargetPrice stockTargetPrice = repository.findByTickerSymbolAndMemberId(request.getTickerSymbol(),
-				memberId)
-			.orElseThrow(() -> new NotFoundResourceException(StockErrorCode.NOT_FOUND_STOCK_TARGET_PRICE));
-		stockTargetPrice.changeIsActive(request.getIsActive());
-		return TargetPriceNotificationUpdateResponse.from(stockTargetPrice);
-	}
-
-	@Secured("ROLE_USER")
-	public TargetPriceNotificationSpecifiedSearchResponse searchTargetPriceNotifications(
+	public TargetPriceNotificationSpecifiedSearchResponse searchStockTargetPrice(
 		String tickerSymbol,
 		Long memberId
 	) {
@@ -236,5 +133,26 @@ public class StockTargetPriceNotificationService {
 		);
 		log.info("특정 종목의 지정가 알림들 조회 결과 : response={}", response);
 		return response;
+	}
+
+	@Transactional
+	@Secured("ROLE_USER")
+	public TargetPriceNotificationUpdateResponse updateStockTargetPrice(
+		TargetPriceNotificationUpdateRequest request, Long memberId) {
+		StockTargetPrice stockTargetPrice = repository.findByTickerSymbolAndMemberId(request.getTickerSymbol(),
+				memberId)
+			.orElseThrow(() -> new NotFoundResourceException(StockErrorCode.NOT_FOUND_STOCK_TARGET_PRICE));
+		stockTargetPrice.changeIsActive(request.getIsActive());
+		return TargetPriceNotificationUpdateResponse.from(stockTargetPrice);
+	}
+
+	// 종목 지정가 단일 제거
+	@Transactional
+	@Authorized(serviceClass = StockTargetPriceAuthorizedService.class)
+	public void deleteStockTargetPrice(@ResourceId Long stockTargetPriceId) {
+		StockTargetPrice stockTargetPrice = repository.findById(stockTargetPriceId)
+			.orElseThrow(() -> new NotFoundResourceException(StockErrorCode.NOT_FOUND_STOCK_TARGET_PRICE));
+		targetPriceNotificationRepository.deleteAllByStockTargetPrices(List.of(stockTargetPrice));
+		repository.deleteById(stockTargetPriceId);
 	}
 }
