+++ src/test/java/codesquad/fineants/domain/stock_target_price/service/StockTargetPriceServiceTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 import org.assertj.core.groups.Tuple;
 import org.junit.jupiter.api.DisplayName;
@@ -13,19 +12,15 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import codesquad.fineants.AbstractContainerBaseTest;
 import codesquad.fineants.domain.common.money.Money;
-import codesquad.fineants.domain.fcm.repository.FcmRepository;
 import codesquad.fineants.domain.kis.domain.dto.response.KisClosingPrice;
 import codesquad.fineants.domain.kis.repository.ClosingPriceRepository;
 import codesquad.fineants.domain.member.domain.entity.Member;
 import codesquad.fineants.domain.member.repository.MemberRepository;
-import codesquad.fineants.domain.notification.repository.NotificationRepository;
-import codesquad.fineants.domain.notificationpreference.repository.NotificationPreferenceRepository;
 import codesquad.fineants.domain.stock.domain.entity.Stock;
 import codesquad.fineants.domain.stock.repository.StockRepository;
 import codesquad.fineants.domain.stock_target_price.domain.dto.request.TargetPriceNotificationCreateRequest;
 import codesquad.fineants.domain.stock_target_price.domain.dto.request.TargetPriceNotificationUpdateRequest;
 import codesquad.fineants.domain.stock_target_price.domain.dto.response.TargetPriceNotificationCreateResponse;
-import codesquad.fineants.domain.stock_target_price.domain.dto.response.TargetPriceNotificationDeleteResponse;
 import codesquad.fineants.domain.stock_target_price.domain.dto.response.TargetPriceNotificationSearchResponse;
 import codesquad.fineants.domain.stock_target_price.domain.dto.response.TargetPriceNotificationSpecifiedSearchResponse;
 import codesquad.fineants.domain.stock_target_price.domain.dto.response.TargetPriceNotificationUpdateResponse;
@@ -39,10 +34,10 @@ import codesquad.fineants.global.errors.exception.BadRequestException;
 import codesquad.fineants.global.errors.exception.FineAntsException;
 import codesquad.fineants.global.errors.exception.NotFoundResourceException;
 
-class StockTargetPriceNotificationServiceTest extends AbstractContainerBaseTest {
+class StockTargetPriceServiceTest extends AbstractContainerBaseTest {
 
 	@Autowired
-	private StockTargetPriceNotificationService service;
+	private StockTargetPriceService service;
 
 	@Autowired
 	private MemberRepository memberRepository;
@@ -57,20 +52,11 @@ class StockTargetPriceNotificationServiceTest extends AbstractContainerBaseTest 
 	private TargetPriceNotificationRepository targetPriceNotificationRepository;
 
 	@Autowired
-	private NotificationRepository notificationRepository;
-
-	@Autowired
-	private FcmRepository fcmRepository;
-
-	@Autowired
-	private NotificationPreferenceRepository notificationPreferenceRepository;
-
-	@Autowired
 	private ClosingPriceRepository manager;
 
 	@DisplayName("사용자는 종목 지정가 알림을 추가합니다")
 	@Test
-	void createStockTargetPriceNotification() {
+	void createStockTargetPrice() {
 		// given
 		Member member = memberRepository.save(createMember());
 		Stock stock = stockRepository.save(createSamsungStock());
@@ -80,7 +66,7 @@ class StockTargetPriceNotificationServiceTest extends AbstractContainerBaseTest 
 			.build();
 
 		// when
-		TargetPriceNotificationCreateResponse response = service.createStockTargetPriceNotification(request,
+		TargetPriceNotificationCreateResponse response = service.createStockTargetPrice(request,
 			member.getId());
 
 		// then
@@ -97,7 +83,7 @@ class StockTargetPriceNotificationServiceTest extends AbstractContainerBaseTest 
 
 	@DisplayName("사용자는 한 종목의 지정가 알림 개수를 5개를 초과할 수 없다")
 	@Test
-	void createStockTargetPriceNotification_whenTargetPriceNotificationLimit_thenThrow400Error() {
+	void createStockTargetPrice_whenTargetPriceNotificationLimit_thenThrow400Error() {
 		// given
 		Member member = memberRepository.save(createMember());
 		Stock stock = stockRepository.save(createSamsungStock());
@@ -112,7 +98,7 @@ class StockTargetPriceNotificationServiceTest extends AbstractContainerBaseTest 
 
 		// when
 		Throwable throwable = catchThrowable(() ->
-			service.createStockTargetPriceNotification(request, member.getId()));
+			service.createStockTargetPrice(request, member.getId()));
 
 		// then
 		assertThat(throwable)
@@ -122,7 +108,7 @@ class StockTargetPriceNotificationServiceTest extends AbstractContainerBaseTest 
 
 	@DisplayName("사용자는 한 종목의 지정가가 이미 존재하는 경우 추가할 수 없다")
 	@Test
-	void createStockTargetPriceNotification_whenExistTargetPrice_thenThrow400Error() {
+	void createStockTargetPrice_whenExistTargetPrice_thenThrow400Error() {
 		// given
 		Member member = memberRepository.save(createMember());
 		Stock stock = stockRepository.save(createSamsungStock());
@@ -135,7 +121,7 @@ class StockTargetPriceNotificationServiceTest extends AbstractContainerBaseTest 
 
 		// when
 		Throwable throwable = catchThrowable(() ->
-			service.createStockTargetPriceNotification(request, member.getId()));
+			service.createStockTargetPrice(request, member.getId()));
 
 		// then
 		assertThat(throwable)
@@ -162,7 +148,7 @@ class StockTargetPriceNotificationServiceTest extends AbstractContainerBaseTest 
 		manager.addPrice(KisClosingPrice.create(stock.getTickerSymbol(), 50000L));
 		manager.addPrice(KisClosingPrice.create(stock2.getTickerSymbol(), 50000L));
 		// when
-		TargetPriceNotificationSearchResponse response = service.searchStockTargetPriceNotification(member.getId());
+		TargetPriceNotificationSearchResponse response = service.searchStockTargetPrices(member.getId());
 
 		// then
 		assertAll(
@@ -208,7 +194,7 @@ class StockTargetPriceNotificationServiceTest extends AbstractContainerBaseTest 
 		manager.addPrice(KisClosingPrice.create(stock.getTickerSymbol(), 50000L));
 
 		// when
-		TargetPriceNotificationSpecifiedSearchResponse response = service.searchTargetPriceNotifications(
+		TargetPriceNotificationSpecifiedSearchResponse response = service.searchStockTargetPrice(
 			stock.getTickerSymbol(), member.getId());
 
 		// then
@@ -231,7 +217,7 @@ class StockTargetPriceNotificationServiceTest extends AbstractContainerBaseTest 
 		Stock stock = stockRepository.save(createSamsungStock());
 
 		// when
-		TargetPriceNotificationSpecifiedSearchResponse response = service.searchTargetPriceNotifications(
+		TargetPriceNotificationSpecifiedSearchResponse response = service.searchStockTargetPrice(
 			stock.getTickerSymbol(), member.getId());
 
 		// then
@@ -260,7 +246,7 @@ class StockTargetPriceNotificationServiceTest extends AbstractContainerBaseTest 
 			.build();
 
 		// when
-		TargetPriceNotificationUpdateResponse response = service.updateStockTargetPriceNotification(
+		TargetPriceNotificationUpdateResponse response = service.updateStockTargetPrice(
 			request, member.getId());
 
 		// then
@@ -293,7 +279,7 @@ class StockTargetPriceNotificationServiceTest extends AbstractContainerBaseTest 
 			.build();
 
 		// when
-		Throwable throwable = catchThrowable(() -> service.updateStockTargetPriceNotification(request, member.getId()));
+		Throwable throwable = catchThrowable(() -> service.updateStockTargetPrice(request, member.getId()));
 
 		// then
 		assertThat(throwable)
@@ -334,196 +320,6 @@ class StockTargetPriceNotificationServiceTest extends AbstractContainerBaseTest 
 		setAuthentication(hacker);
 		// when
 		Throwable throwable = catchThrowable(() -> service.deleteStockTargetPrice(stockTargetPrice.getId()));
-		// then
-		assertThat(throwable)
-			.isInstanceOf(FineAntsException.class)
-			.hasMessage(MemberErrorCode.FORBIDDEN_MEMBER.getMessage());
-	}
-
-	@DisplayName("사용자는 종목 지정가 알림을 제거합니다")
-	@Test
-	void deleteStockTargetPriceNotification() {
-		// given
-		Member member = memberRepository.save(createMember());
-		Stock stock = stockRepository.save(createSamsungStock());
-		StockTargetPrice stockTargetPrice = repository.save(createStockTargetPrice(member, stock));
-		TargetPriceNotification targetPriceNotification = targetPriceNotificationRepository.save(
-			createTargetPriceNotification(stockTargetPrice));
-
-		Long id = targetPriceNotification.getId();
-
-		setAuthentication(member);
-		// when
-		TargetPriceNotificationDeleteResponse response = service.deleteStockTargetPriceNotification(id);
-
-		// then
-		assertAll(
-			() -> assertThat(response)
-				.extracting("deletedIds")
-				.asList()
-				.hasSize(1),
-			() -> assertThat(targetPriceNotificationRepository.findById(id).isEmpty()).isTrue(),
-			() -> assertThat(repository.findById(stockTargetPrice.getId()).isEmpty()).isTrue()
-		);
-	}
-
-	@DisplayName("사용자는 존재하지 않는 지정가를 삭제할 수 없습니다")
-	@Test
-	void deleteStockTargetPriceNotification_whenNotExistTargetPriceNotificationIds_thenThrow404Error() {
-		// given
-		Member member = memberRepository.save(createMember());
-		Stock stock = stockRepository.save(createSamsungStock());
-		StockTargetPrice stockTargetPrice = repository.save(createStockTargetPrice(member, stock));
-		targetPriceNotificationRepository.save(createTargetPriceNotification(stockTargetPrice));
-
-		Long id = 9999L;
-
-		setAuthentication(member);
-		// when
-		Throwable throwable = catchThrowable(() -> service.deleteStockTargetPriceNotification(id));
-
-		// then
-		assertThat(throwable)
-			.isInstanceOf(NotFoundResourceException.class)
-			.hasMessage(StockErrorCode.NOT_FOUND_TARGET_PRICE.getMessage());
-	}
-
-	@DisplayName("사용자는 다른 사용자의 지정가 알림을 삭제할 수 없습니다.")
-	@Test
-	void deleteStockTargetPriceNotification_whenForbiddenTargetPriceNotificationIds_thenThrow403Error() {
-		// given
-		Member member = memberRepository.save(createMember());
-		Stock stock = stockRepository.save(createSamsungStock());
-		StockTargetPrice stockTargetPrice = repository.save(createStockTargetPrice(member, stock));
-		TargetPriceNotification targetPriceNotification = targetPriceNotificationRepository.save(
-			createTargetPriceNotification(stockTargetPrice));
-
-		Member hacker = memberRepository.save(createMember("일개미4567", "kim2@gmail.com"));
-
-		setAuthentication(hacker);
-		// when
-		Throwable throwable = catchThrowable(
-			() -> service.deleteStockTargetPriceNotification(targetPriceNotification.getId()));
-
-		// then
-		assertThat(throwable)
-			.isInstanceOf(FineAntsException.class)
-			.hasMessage(MemberErrorCode.FORBIDDEN_MEMBER.getMessage());
-	}
-
-	@DisplayName("사용자는 종목 지정가 알림 전체를 삭제합니다")
-	@Test
-	void deleteAllStockTargetPriceNotification() {
-		// given
-		Member member = memberRepository.save(createMember());
-		Stock stock = stockRepository.save(createSamsungStock());
-		StockTargetPrice stockTargetPrice = repository.save(createStockTargetPrice(member, stock));
-		List<TargetPriceNotification> targetPriceNotifications = targetPriceNotificationRepository.saveAll(
-			createTargetPriceNotification(stockTargetPrice, List.of(60000L, 70000L)));
-
-		List<Long> ids = targetPriceNotifications.stream()
-			.map(TargetPriceNotification::getId)
-			.collect(Collectors.toList());
-
-		setAuthentication(member);
-		// when
-		TargetPriceNotificationDeleteResponse response = service.deleteAllStockTargetPriceNotification(
-			ids,
-			stock.getTickerSymbol(),
-			member.getId()
-		);
-
-		// then
-		assertAll(
-			() -> assertThat(response)
-				.extracting("deletedIds")
-				.asList()
-				.hasSize(2),
-			() -> assertThat(targetPriceNotificationRepository.findAllById(ids).isEmpty()).isTrue(),
-			() -> assertThat(repository.findById(stockTargetPrice.getId()).isEmpty()).isTrue()
-		);
-	}
-
-	@DisplayName("사용자는 종목 지정가 알림을 전체 삭제할 때, 존재하지 않는 지정가 알림을 제거할 수 없습니다.")
-	@Test
-	void deleteAllStockTargetPriceNotification_whenNotExistTargetPriceNotificationIds_thenThrow404Error() {
-		// given
-		Member member = memberRepository.save(createMember());
-		Stock stock = stockRepository.save(createSamsungStock());
-		StockTargetPrice stockTargetPrice = repository.save(createStockTargetPrice(member, stock));
-		List<TargetPriceNotification> targetPriceNotifications = targetPriceNotificationRepository.saveAll(
-			createTargetPriceNotification(stockTargetPrice, List.of(60000L, 70000L)));
-
-		List<Long> ids = targetPriceNotifications.stream()
-			.map(TargetPriceNotification::getId)
-			.collect(Collectors.toList());
-		ids.add(9999L);
-
-		setAuthentication(member);
-		// when
-		Throwable throwable = catchThrowable(() -> service.deleteAllStockTargetPriceNotification(
-			ids,
-			stock.getTickerSymbol(),
-			member.getId()
-		));
-
-		// then
-		assertThat(throwable)
-			.isInstanceOf(NotFoundResourceException.class)
-			.hasMessage(StockErrorCode.NOT_FOUND_TARGET_PRICE.getMessage());
-	}
-
-	@DisplayName("사용자는 종목 지정가 알림을 전체 삭제할 때, 존재하지 않는 종목에 대해서 제거할 수 없습니다.")
-	@Test
-	void deleteAllStockTargetPriceNotification_whenNotExistStock_thenThrow404Error() {
-		// given
-		Member member = memberRepository.save(createMember());
-		Stock stock = stockRepository.save(createSamsungStock());
-		StockTargetPrice stockTargetPrice = repository.save(createStockTargetPrice(member, stock));
-		List<TargetPriceNotification> targetPriceNotifications = targetPriceNotificationRepository.saveAll(
-			createTargetPriceNotification(stockTargetPrice, List.of(60000L, 70000L)));
-
-		List<Long> ids = targetPriceNotifications.stream()
-			.map(TargetPriceNotification::getId)
-			.collect(Collectors.toList());
-
-		setAuthentication(member);
-		// when
-		Throwable throwable = catchThrowable(() -> service.deleteAllStockTargetPriceNotification(
-			ids,
-			"999999",
-			member.getId()
-		));
-
-		// then
-		assertThat(throwable)
-			.isInstanceOf(NotFoundResourceException.class)
-			.hasMessage(StockErrorCode.NOT_FOUND_STOCK.getMessage());
-	}
-
-	@DisplayName("사용자는 종목 지정가 알림을 전체 삭제할 때, 다른 사용자의 지정가 알림을 삭제할 수 없습니다")
-	@Test
-	void deleteAllStockTargetPriceNotification_whenForbiddenTargetPriceNotifications_thenThrow403Error() {
-		// given
-		Member member = memberRepository.save(createMember());
-		Member hacker = memberRepository.save(createMember("hacker"));
-		Stock stock = stockRepository.save(createSamsungStock());
-		StockTargetPrice stockTargetPrice = repository.save(createStockTargetPrice(member, stock));
-		List<TargetPriceNotification> targetPriceNotifications = targetPriceNotificationRepository.saveAll(
-			createTargetPriceNotification(stockTargetPrice, List.of(60000L, 70000L)));
-
-		List<Long> ids = targetPriceNotifications.stream()
-			.map(TargetPriceNotification::getId)
-			.collect(Collectors.toList());
-
-		setAuthentication(hacker);
-		// when
-		Throwable throwable = catchThrowable(() -> service.deleteAllStockTargetPriceNotification(
-			ids,
-			stock.getTickerSymbol(),
-			hacker.getId()
-		));
-
 		// then
 		assertThat(throwable)
 			.isInstanceOf(FineAntsException.class)
